Move App Router client-side constants to separate file
