+++ packages/next/src/client/components/redirect.ts
@@ -1,7 +1,7 @@
 import { requestAsyncStorage } from './request-async-storage.external'
 import type { ResponseCookies } from '../../server/web/spec-extension/cookies'
 import { actionAsyncStorage } from './action-async-storage.external'
-import { RedirectStatusCode } from '../../shared/lib/constants'
+import { RedirectStatusCode } from '../../shared/lib/app-router-client-constants'
 
 const REDIRECT_ERROR_CODE = 'NEXT_REDIRECT'
