+++ packages/next/src/lib/constants.ts
@@ -6,8 +6,6 @@ export const PRERENDER_REVALIDATE_HEADER = 'x-prerender-revalidate'
 export const PRERENDER_REVALIDATE_ONLY_GENERATED_HEADER =
   'x-prerender-revalidate-if-generated'
 
-export const NEXT_DID_POSTPONE_HEADER = 'x-nextjs-postponed'
-
 export const RSC_PREFETCH_SUFFIX = '.prefetch.rsc'
 export const RSC_SUFFIX = '.rsc'
 export const NEXT_DATA_SUFFIX = '.json'
