+++ packages/next/src/shared/lib/app-router-client-constants.ts
@@ -1,0 +1,6 @@
+export enum RedirectStatusCode {
+  SeeOther = 303,
+  TemporaryRedirect = 307,
+  PermanentRedirect = 308,
+}
+export const NEXT_DID_POSTPONE_HEADER = 'x-nextjs-postponed'
